changed an int to a float

+++ Source/EBSGFramework/EBSGExtension.cs
@@ -145,7 +145,7 @@ namespace EBSGFramework
         public ThingDef chargeMote;
         public ThingDef chargeMotePulse;
         public EffecterDef wasteProducedEffecter;
-        public int wastePerHourOfUse; // Only applies if there's a WasteProducer comp
+        public float wastePerHourOfUse; // Only applies if there's a WasteProducer comp
         public List<GeneEffect> resourceOffsetsPerHour; // For DRG in the pawn need charger
         public bool negativeNeedOffsetsAreNotCosts = false;
         public bool negativeResourceOffsetsAreNotCosts = false;
